backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
psycopg2-binary
